+++ packages/sdk-react/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @turnkey/sdk-react
 
+## 5.4.9
+
+### Patch Changes
+
+- Updated dependencies [[`71cdca3`](https://github.com/tkhq/sdk/commit/71cdca3b97ba520dc5327410a1e82cf9ad85fb0e), [`9fbd5c4`](https://github.com/tkhq/sdk/commit/9fbd5c459782dc3721dd0935d0a4458babce258b)]:
+  - @turnkey/sdk-server@4.11.0
+  - @turnkey/sdk-browser@5.12.0
+  - @turnkey/crypto@2.8.4
+  - @turnkey/wallet-stamper@1.1.6
+
 ## 5.4.8
 
 ### Patch Changes